Use repairwheel to correct references between whisper libraries.
I'm still unsure if this is the best approach but I wanted `pip install .` to  work, normally there is an option in cibuildwheel that let us run wheelrepair, but this leaves the pip install broken.

It is quite difficult to find information how to do it properly, so I'm still unsure that it is right solution as we are still ending up with libs in two places in the wheel, but at least they reference them selfs correctly so which ever is picked first the rest is loaded.

I've tested this only on my mac so I'm not sure that it is working correctly in windows and linux but it should as repairwheel is multiplatform.

After repair the libs are in pywhispercpp-1.2.0/*.dylib and in pywhispercpp-1.2.0/pywhispercpp/.dylibs/*.dylib, but they reference each other have a look at the output of otool -L
```
otool -L pywhispercpp-1.2.0/libwhisper.1.7.1.dylib
pywhispercpp-1.2.0/libwhisper.1.7.1.dylib:
        @rpath/libwhisper.1.dylib (compatibility version 1.0.0, current version 1.7.1)
        @loader_path/pywhispercpp/.dylibs/libggml.dylib (compatibility version 0.0.0, current version 0.0.0)
        @loader_path/pywhispercpp/.dylibs/libwhisper.coreml.dylib (compatibility version 0.0.0, current version 0.0.0)
        /usr/lib/libc++.1.dylib (compatibility version 1.0.0, current version 1800.101.0)
        /usr/lib/libSystem.B.dylib (compatibility version 1.0.0, current version 1351.0.0)
```

````
otool -L pywhispercpp-1.2.0/pywhispercpp/.dylibs/libwhisper.1.7.1.dylib
pywhispercpp-1.2.0/pywhispercpp/.dylibs/libwhisper.1.7.1.dylib:
        /DLC/pywhispercpp/.dylibs/libwhisper.1.7.1.dylib (compatibility version 1.0.0, current version 1.7.1)
        @loader_path/libggml.dylib (compatibility version 0.0.0, current version 0.0.0)
        @loader_path/libwhisper.coreml.dylib (compatibility version 0.0.0, current version 0.0.0)
        /usr/lib/libc++.1.dylib (compatibility version 1.0.0, current version 1800.101.0)
        /usr/lib/libSystem.B.dylib (compatibility version 1.0.0, current version 1351.0.0)
```